Serialization order in class Version Moved class Version to model package
